Begin adding fam member related functionality
Mainly adding family members to tree. Not functional yet.
State management starts becoming clunky, due to all the threading
of props downstream in many directions from the top of the tree.
Through many levels of the tree, and then handling events back up.

Might branch out into Redux or Context because of that.