Add many changes, read rest of msg
Added zooming in and out with scaling of fam members through
special css classes for different zoom levels.

Added rendering of lines connecting fam members using new components,
with some processing of units, offsets, scaling and stuff regarding this
and the zoom feature. The component that holds the lines connecting
members also acts as a stretcher of the view, so dragging past the
bounds of the tree is easier. This app this way is approaching
minimum viable feature set status.