Fix annoying <select> problem
It turns out that the text displayed by the select tag is not the value
of the select, but is the inner text of the option that has matching
value to the select tag's value.

Eg if the select tag has value = 3, it will try to find the option that
has value 3 and take the inner text from there, as opposed to just
displaying 3.

This was causing confusion and it seemed as if a bug existed,
where in fact it did not, the approach was just based on wrong info,
so some other things were not working.