Add family shifting
When a member's location is below a set minimum (x:1, y:1) as of now),
the whole tree will shift to make room for it, so that it never goes
out of view. This is a solution for the fact that absolutely positioned
elements that have top or left in the negative don't cause scrolling
on the parent element, but they do in the positive.