Add better validation for date input
The previous single source of truth validation was too restrictive on
the user, since it hard blocked typing, eg. when the user tried to
erase the n of day to then type a new one, it wouldn't let them erase,
since a "" is not a valid day of month. Now the input maintains a
secondary source of truth with temporary, unvalidated values, so they
can type and see that they are invalid and only when they are valid,
they get transferred to the primary source.